Crate massa_protocol_exports

protocol component exports

Modules

channels 🔒

error 🔒

protocol_controller 🔒

settings 🔒

tests

TODO: Add only if test. Removed the configuration test because don’t work if running cargo test on an other sub-crate.

Structs

ProtocolCommandSender

protocol command sender

ProtocolConfig

Dynamic protocol configuration mix in static settings and constants configurations.

ProtocolManager

protocol manager used to stop the protocol

ProtocolReceivers

Contains channels(receivers) used by the protocol worker

ProtocolSenders

Contains channels (senders) used by the protocol worker Contains (a) channel(s) to send info to api

Enums

ProtocolCommand

Commands that protocol worker can process

ProtocolError

protocol error

ProtocolManagementCommand

protocol management commands

Type Definitions

BlocksResults

block result: map block id to