Enum massa_protocol_exports::ProtocolError

#[non_exhaustive]
pub enum ProtocolError {

    WrongSignature,
    GeneralProtocolError(String),
    ChannelError(String),
    TokioTaskJoinError(JoinError),
    TokieRecvError(RecvError),
    TokioSendError(SendError<ProtocolEvent>),
    PeerConnectionError(NetworkConnectionErrorType),
    InvalidIpError(IpAddr),
    ActiveConnectionMissing(ConnectionId),
    IOError(Error),
    SerdeError(Error),
    MassaHashError(MassaHashError),
    HandshakeError(HandshakeErrorType),
    UnexpectedNodeCommandChannelClosure,
    UnexpectedWriterClosure,
    TimeError(TimeError),
    MissingPeersError,
    ModelsError(ModelsError),
    NetworkError(NetworkError),
    ContainerInconsistencyError(String),

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

WrongSignature

wrong signature

GeneralProtocolError(String)

Tuple Fields

0: String

Protocol err:{0}

ChannelError(String)

Tuple Fields

0: String

An error occurred during channel communication: {0}

TokioTaskJoinError(JoinError)

Tuple Fields

0: JoinError

A tokio task has crashed err:{0}

TokieRecvError(RecvError)

Tuple Fields

0: RecvError

error receiving oneshot response : {0}

TokioSendError(SendError<ProtocolEvent>)

Tuple Fields

0: SendError<ProtocolEvent>

error sending protocol event: {0}

PeerConnectionError(NetworkConnectionErrorType)

Tuple Fields

0: NetworkConnectionErrorType

Error during network connection:{0:?}

InvalidIpError(IpAddr)

Tuple Fields

0: IpAddr

The ip:{0} address is not valid

ActiveConnectionMissing(ConnectionId)

Tuple Fields

0: ConnectionId

Active connection missing:{0}

IOError(Error)

Tuple Fields

0: Error

IO error : {0}

SerdeError(Error)

Tuple Fields

0: Error

Serde error : {0}

MassaHashError(MassaHashError)

Tuple Fields

0: MassaHashError

massa_hash error {0}

HandshakeError(HandshakeErrorType)

Tuple Fields

0: HandshakeErrorType

handshake error:{0:?}

UnexpectedNodeCommandChannelClosure

the network controller should not drop a node command sender before shutting down the node.

UnexpectedWriterClosure

the writer of a node should not drop its event sender before sending a clean_exit message.

TimeError(TimeError)

Tuple Fields

0: TimeError

Time error {0}

MissingPeersError

missing peers

ModelsError(ModelsError)

Tuple Fields

0: ModelsError

models error: {0}

NetworkError(NetworkError)

Tuple Fields

0: NetworkError

network error: {0}

ContainerInconsistencyError(String)

Tuple Fields

0: String

container inconsistency error: {0}