Enum massa_protocol_exports::error::ProtocolError

#[non_exhaustive]
pub enum ProtocolError {

    WrongSignature,
    GeneralProtocolError(String),
    ChannelError(String),
    TokioTaskJoinError(JoinError),
    TokioRecvError(RecvError),
    PeerConnectionError(NetworkConnectionErrorType),
    InvalidIpError(IpAddr),
    ActiveConnectionMissing(ConnectionId),
    IOError(Error),
    SerdeError(Error),
    MassaHashError(MassaHashError),
    UnexpectedNodeCommandChannelClosure,
    UnexpectedWriterClosure,
    TimeError(TimeError),
    MissingPeersError,
    ModelsError(ModelsError),
    NetworkError(NetworkError),
    ContainerInconsistencyError(String),
    InvalidOperationError(String),

}

protocol error

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

WrongSignature

Wrong signature

GeneralProtocolError(String)

Protocol error: {0}

ChannelError(String)

An error occurred during channel communication: {0}

TokioTaskJoinError(JoinError)

A tokio task has crashed err: {0}

TokioRecvError(RecvError)

Error receiving one shot response: {0}

PeerConnectionError(NetworkConnectionErrorType)

Error during network connection: {0:?}

InvalidIpError(IpAddr)

The ip: {0} address is not valid

ActiveConnectionMissing(ConnectionId)

Active connection missing: {0}

IOError(Error)

IO error: {0}

SerdeError(Error)

Serde error: {0}

MassaHashError(MassaHashError)

massa_hash error: {0}

UnexpectedNodeCommandChannelClosure

The network controller should not drop a node command sender before shutting down the node.

UnexpectedWriterClosure

The writer of a node should not drop its event sender before sending a clean_exit message.

TimeError(TimeError)

Time error: {0}

MissingPeersError

Missing peers

ModelsError(ModelsError)

Models error: {0}

NetworkError(NetworkError)

Network error: {0}

ContainerInconsistencyError(String)

Container inconsistency error: {0}

InvalidOperationError(String)

Invalid operation error: {0}