Struct massa_protocol_exports::settings::ProtocolConfig

pub struct ProtocolConfig {

    pub keypair_file: PathBuf,
    pub listeners: HashMap<SocketAddr, TransportType>,
    pub initial_peers: PathBuf,
    pub ask_block_timeout: MassaTime,
    pub max_blocks_kept_for_propagation: usize,
    pub max_block_propagation_time: MassaTime,
    pub block_propagation_tick: MassaTime,
    pub max_known_blocks_size: usize,
    pub max_node_known_blocks_size: usize,
    pub max_node_wanted_blocks_size: usize,
    pub max_known_ops_size: usize,
    pub max_node_known_ops_size: usize,
    pub max_known_endorsements_size: usize,
    pub max_node_known_endorsements_size: usize,
    pub max_simultaneous_ask_blocks_per_node: usize,
    pub max_send_wait: MassaTime,
    pub operation_batch_buffer_capacity: usize,
    pub operation_announcement_buffer_capacity: usize,
    pub operation_batch_proc_period: MassaTime,
    pub asked_operations_buffer_capacity: usize,
    pub operation_announcement_interval: MassaTime,
    pub max_operation_storage_time: MassaTime,
    pub max_operations_per_message: u64,
    pub max_operations_per_block: u32,
    pub max_serialized_operations_size_per_block: usize,
    pub controller_channel_size: usize,
    pub event_channel_size: usize,
    pub t0: MassaTime,
    pub genesis_timestamp: MassaTime,
    pub max_ops_kept_for_propagation: usize,
    pub max_operations_propagation_time: MassaTime,
    pub max_endorsements_propagation_time: MassaTime,
    pub max_message_size: usize,
    pub thread_tester_count: u8,
    pub max_size_channel_commands_connectivity: usize,
    pub max_size_channel_commands_retrieval_operations: usize,
    pub max_size_channel_commands_propagation_operations: usize,
    pub max_size_channel_commands_retrieval_endorsements: usize,
    pub max_size_channel_commands_propagation_endorsements: usize,
    pub max_size_channel_commands_retrieval_blocks: usize,
    pub max_size_channel_commands_propagation_blocks: usize,
    pub max_size_channel_commands_peers: usize,
    pub max_size_channel_commands_peer_testers: usize,
    pub max_size_channel_network_to_operation_handler: usize,
    pub max_size_channel_network_to_block_handler: usize,
    pub max_size_channel_network_to_endorsement_handler: usize,
    pub max_size_channel_network_to_peer_handler: usize,
    pub endorsement_count: u32,
    pub thread_count: u8,
    pub max_size_value_datastore: u64,
    pub max_size_function_name: u16,
    pub max_size_call_sc_parameter: u32,
    pub max_op_datastore_entry_count: u64,
    pub max_op_datastore_key_length: u8,
    pub max_op_datastore_value_length: u64,
    pub max_denunciations_in_block_header: u32,
    pub max_endorsements_per_message: u64,
    pub max_size_peers_announcement: u64,
    pub max_size_listeners_per_peer: u64,
    pub last_start_period: u64,
    pub try_connection_timer: MassaTime,
    pub try_connection_timer_same_peer: MassaTime,
    pub unban_everyone_timer: MassaTime,
    pub max_in_connections: usize,
    pub timeout_connection: MassaTime,
    pub message_timeout: MassaTime,
    pub tester_timeout: MassaTime,
    pub read_write_limit_bytes_per_second: u128,
    pub routable_ip: Option<IpAddr>,
    pub debug: bool,
    pub peers_categories: HashMap<String, PeerCategoryInfo>,
    pub default_category_info: PeerCategoryInfo,
    pub version: Version,
    pub test_oldest_peer_cooldown: MassaTime,

}

Dynamic protocol configuration mix in static settings and constants configurations.

Fields
keypair_file: PathBuf

self keypair

listeners: HashMap<SocketAddr, TransportType>

listeners from where we can receive messages

initial_peers: PathBuf

initial peers path

ask_block_timeout: MassaTime

after ask_block_timeout milliseconds we try to ask a block to another node

max_blocks_kept_for_propagation: usize

Max known blocks we keep during their propagation

max_block_propagation_time: MassaTime

Time during which a block is expected to propagate

block_propagation_tick: MassaTime

Block propagation tick interval, useful for propagating blocks quickly to newly connected peers.

max_known_blocks_size: usize

max known blocks of current nodes we keep in memory

max_node_known_blocks_size: usize

max known blocks of foreign nodes we keep in memory (by node)

max_node_wanted_blocks_size: usize

max wanted blocks per node kept in memory

max_known_ops_size: usize

max known operations current node kept in memory

max_node_known_ops_size: usize

max known operations of foreign nodes we keep in memory (by node)

max_known_endorsements_size: usize

max known endorsements by our node that we kept in memory

max_node_known_endorsements_size: usize

max known endorsements of foreign nodes we keep in memory (by node)

max_simultaneous_ask_blocks_per_node: usize

we ask for the same block max_simultaneous_ask_blocks_per_node times at the same time

max_send_wait: MassaTime

Max wait time for sending a Network or Node event.

operation_batch_buffer_capacity: usize

Maximum number of batches in the memory buffer. Dismiss the new batches if overflow

operation_announcement_buffer_capacity: usize

Maximum number of operations in the announcement buffer. Immediately announce if overflow.

operation_batch_proc_period: MassaTime

Start processing batches in the buffer each operation_batch_proc_period in millisecond

asked_operations_buffer_capacity: usize

Maximum number of asked operations in the memory buffer.

operation_announcement_interval: MassaTime

Interval at which operations are announced in batches.

max_operation_storage_time: MassaTime

Maximum time we keep an operation in the storage

max_operations_per_message: u64

Maximum of operations sent in one message.

max_operations_per_block: u32

Maximum of operations sent in one block.

max_serialized_operations_size_per_block: usize

Maximum size in bytes of all serialized operations size in a block

controller_channel_size: usize

Controller channel size

event_channel_size: usize

Event channel size

t0: MassaTime

t0

genesis_timestamp: MassaTime

Genesis timestamp

max_ops_kept_for_propagation: usize

max number of operations kept in memory for propagation

max_operations_propagation_time: MassaTime

max time we propagate operations

max_endorsements_propagation_time: MassaTime

max time we propagate endorsements

max_message_size: usize

Max message size

thread_tester_count: u8

number of thread tester

max_size_channel_commands_connectivity: usize

Max size of the channel for command to the connectivity thread

max_size_channel_commands_retrieval_operations: usize

Max size of channel to send commands to retrieval thread of operations

max_size_channel_commands_propagation_operations: usize

Max size of channel to send commands to propagation thread of operations

max_size_channel_commands_retrieval_endorsements: usize

Max size of channel to send commands to retrieval thread of endorsements

max_size_channel_commands_propagation_endorsements: usize

Max size of channel to send commands to propagation thread of blocks

max_size_channel_commands_retrieval_blocks: usize

Max size of channel to send commands to retrieval thread of blocks

max_size_channel_commands_propagation_blocks: usize

Max size of channel to send commands to propagation thread of blocks

max_size_channel_commands_peers: usize

Max size of channel to send commands to thread of peers

max_size_channel_commands_peer_testers: usize

Max size of channel to send commands to the peer testers

max_size_channel_network_to_operation_handler: usize

Max size of channel that transfer message from network to operation handler

max_size_channel_network_to_block_handler: usize

Max size of channel that transfer message from network to block handler

max_size_channel_network_to_endorsement_handler: usize

Max size of channel that transfer message from network to endorsement handler

max_size_channel_network_to_peer_handler: usize

Max size of channel that transfer message from network to peer handler

endorsement_count: u32

endorsements per block

thread_count: u8

running threads count

max_size_value_datastore: u64

Maximum size of an value user datastore

max_size_function_name: u16

Maximum size of a function name

max_size_call_sc_parameter: u32

Maximum size of a parameter of a call in ops

max_op_datastore_entry_count: u64max_op_datastore_key_length: u8max_op_datastore_value_length: u64max_denunciations_in_block_header: u32

Maximum number of denunciations in a block header

max_endorsements_per_message: u64

Maximum number of endorsements that can be propagated in one message

max_size_peers_announcement: u64

Maximum number of peers per announcement

max_size_listeners_per_peer: u64

Maximum number of listeners per peer

last_start_period: u64

Last start period

try_connection_timer: MassaTime

try connection timer

try_connection_timer_same_peer: MassaTime

try connection timer same peer

unban_everyone_timer: MassaTime

periodically unban every peer

max_in_connections: usize

Max in connections

timeout_connection: MassaTime

Timeout connection

message_timeout: MassaTime

Timeout message

tester_timeout: MassaTime

Timeout for the tester operations

read_write_limit_bytes_per_second: u128

Number of bytes per second that can be read/write in a connection (should be a 10 multiplier)

routable_ip: Option<IpAddr>

Optional routable ip

debug: bool

debug prints

peers_categories: HashMap<String, PeerCategoryInfo>

Peers categories infos

default_category_info: PeerCategoryInfo

Default category infos

version: Version

Version

test_oldest_peer_cooldown: MassaTime

Cooldown before testing again an old peer