Module massa_protocol_exports::tests::tools

test utilities

Structs

NodeInfo

test utility structures keeps keypair and associated node id

PROTOCOL_SETTINGS

protocol settings

Functions

asked_list

retrieve what blocks where asked to which nodes

assert_banned_nodes

assert a list of node(s) has been banned

assert_hash_asked_to_node

assert block id has been asked to node

create_and_connect_nodes

create number of nodes and connect them with protocol

create_block

Creates a block for use in protocol, without paying attention to consensus related things like slot, parents, and merkle root.

create_block_with_endorsements

create a block with no operation

create_block_with_operations

create a block with no endorsement

create_endorsement

Creates an endorsement for use in protocol tests, without paying attention to consensus related things.

create_node

create node info

create_operation_with_expire_period

Create an operation, from a specific sender, and with a specific expire period.

create_protocol_settings

create a ProtocolConfig with typical values

send_and_propagate_block

send a block and assert it has been propagate (or not)

wait_protocol_event

wait protocol event

wait_protocol_pool_event

wait protocol pool event